<FIX> JwtUtils 관련 수정 (멤버 변수 및 메서드)
