+++ src/main/java/roomescape/authentication/AuthAdminRoleInterceptor.java
@@ -5,7 +5,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.authentication.jwt.JwtProvider;
 import roomescape.authentication.jwt.JwtResponse;
 import roomescape.authentication.jwt.JwtUtils;
 
@@ -13,11 +12,11 @@ import roomescape.authentication.jwt.JwtUtils;
 @RequiredArgsConstructor
 public class AuthAdminRoleInterceptor implements HandlerInterceptor {
 
-    private final JwtProvider jwtProvider;
+    private final JwtUtils jwtUtils;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        JwtResponse jwtResponse = JwtUtils.extractTokenFromCookie(request.getCookies());
+        JwtResponse jwtResponse = jwtUtils.extractTokenFromCookie(request.getCookies());
         if (jwtResponse.accessToken() == null || !isAdmin(jwtResponse.accessToken())) {
             response.setStatus(401);
             return false;
@@ -27,7 +26,7 @@ public class AuthAdminRoleInterceptor implements HandlerInterceptor {
     }
 
     private boolean isAdmin(String token) {
-        MemberAuthInfo memberAuthInfo = JwtUtils.extractMemberAuthInfoFromToken(token);
+        MemberAuthInfo memberAuthInfo = jwtUtils.extractMemberAuthInfoFromToken(token);
         return "ADMIN".equals(memberAuthInfo.role());
     }
 }
