+++ src/main/java/roomescape/authentication/jwt/JwtUtils.java
@@ -9,7 +9,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.authentication.MemberAuthInfo;
+import roomescape.exception.JwtProviderException;
 import roomescape.exception.JwtValidationException;
+import roomescape.member.Member;
 
 import java.util.Arrays;
 
@@ -17,14 +19,25 @@ import java.util.Arrays;
 @RequiredArgsConstructor
 public class JwtUtils {
 
-    private static String secretKey;
-
     @Value("${roomescape.auth.jwt.secret}")
-    public void setSecretKey(String secretKey) {
-        JwtUtils.secretKey = secretKey;
+    private String secretKey;
+
+    public JwtResponse createAccessToken(Member member) {
+        try {
+            String accessToken = Jwts.builder()
+                    .setSubject(member.getId().toString())
+                    .claim("name", member.getName())
+                    .claim("role", member.getRole())
+                    .signWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
+                    .compact();
+
+            return new JwtResponse(accessToken);
+        } catch (JwtException e) {
+            throw new JwtProviderException("JWT 생성에 실패하였습니다.");
+        }
     }
 
-    public static MemberAuthInfo extractMemberAuthInfoFromToken(String token) {
+    public MemberAuthInfo extractMemberAuthInfoFromToken(String token) {
         if (token == null || token.isEmpty()) {
             throw new JwtValidationException("토큰이 존재하지 않습니다.");
         }
@@ -45,7 +58,7 @@ public class JwtUtils {
         }
     }
 
-    public static JwtResponse extractTokenFromCookie(Cookie[] cookies) {
+    public JwtResponse extractTokenFromCookie(Cookie[] cookies) {
         if (cookies == null) {
             throw new JwtValidationException("쿠키가 존재하지 않습니다.");
         }
